Move new NexusOutputsMgr classes.
Move headers for class and subclasses from include/utilities/ to
include/utilities/output/.
